chore: rename `GadgetCaller` to `BlackBoxFuncCaller`
