+++ barretenberg_static_lib/src/acvm_interop/pwg/black_box_functions.rs
@@ -2,13 +2,13 @@ use crate::Barretenberg;
 
 use common::acvm::acir::{circuit::opcodes::BlackBoxFuncCall, native_types::Witness};
 use common::acvm::{FieldElement, OpcodeResolutionError};
-use common::gadget_caller::BarretenbergShared;
+use common::black_box_functions::BarretenbergShared;
 use std::collections::BTreeMap;
 
 // Note that the outputs for things like Sha256 need to be computed
 // as they may be used in later arithmetic gates
 
-pub(super) struct GadgetCaller;
+pub(super) struct BlackBoxFuncCaller;
 
 impl BarretenbergShared for Barretenberg {
     fn new() -> Self {
@@ -33,12 +33,12 @@ impl BarretenbergShared for Barretenberg {
     }
 }
 
-impl GadgetCaller {
+impl BlackBoxFuncCaller {
     pub(super) fn solve_black_box_func_call(
         initial_witness: &mut BTreeMap<Witness, FieldElement>,
         gadget_call: &BlackBoxFuncCall,
     ) -> Result<(), OpcodeResolutionError> {
-        common::gadget_caller::solve_black_box_func_call::<Barretenberg>(
+        common::black_box_functions::solve_black_box_func_call::<Barretenberg>(
             initial_witness,
             gadget_call,
         )
