+++ common/src/black_box_functions.rs
