+++ common/src/lib.rs
@@ -2,8 +2,8 @@
 #![warn(unreachable_pub)]
 
 pub mod barretenberg_structures;
+pub mod black_box_functions;
 pub mod contract;
-pub mod gadget_caller;
 
 #[cfg(feature = "std")]
 pub mod crs;
